+++ cg/store/api/findbusinessdata.py
@@ -132,6 +132,17 @@ class FindBusinessDataHandler(BaseHandler):
             .all()
         )
 
+    def get_samples_by_family_id(self, family_id: str) -> List[models.Sample]:
+        """Get samples on a given family_id"""
+        case: models.Family = self.family(internal_id=family_id)
+        return [link.sample for link in case.links] if case else []
+
+    def get_sequenced_samples(self, family_id: str) -> List[models.Sample]:
+        """Get sequenced samples by family_id"""
+
+        samples: List[models.Sample] = self.get_samples_by_family_id(family_id)
+        return [sample for sample in samples if sample.sequencing_qc]
+
     def find_family(self, customer: models.Customer, name: str) -> models.Family:
         """Find a family by family name within a customer."""
         return self.Family.query.filter_by(customer=customer, name=name).first()